Objeto para Comunidades segmentado en Componente
