+++ react/src/components/consultaConMapa/graficosDinamicos.ts
@@ -1,3 +1,6 @@
+import logger from 'utilidades/logger';
+import dynamic from 'next/dynamic';
+
 import * as turf from '@turf/turf';
 import { Circle, Layer, Path } from 'leaflet';
 
@@ -15,8 +18,6 @@ import {
   FeatureTerritorios, 
   FeatureLineas
 } from './tipos';
-
-import logger from 'utilidades/logger';
 
 const colorParaLineasColindantes: { [key: string]: string } = {};
